requirements.txt
fastapi==0.115.2
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.35
uvicorn==0.31.1
cryptography==43.0.1
PyJWT==2.9.0
python-multipart==0.0.16
Werkzeug==3.0.4
pydantic-settings==2.5.2
# alembic==1.14.0
pytz==2024.2
websockets==14.1
google-auth==2.37.0
google-auth-httplib2==0.2.0
requests==2.32.3